Clarify setup a little
Added some hints that made it easier (at least for me) to get this running. Even though you can be flexible with the naming, it's surely easier for someone setting this up to work with a consistent schema for the host names since the nomenclature is difficult enough, so I renamed everything to `demosite[-origin].dev`.

I hope this helps a bit.